xapi/info: replace add with push/unshift/insert
Enables controlling the order in which infos are rendered to a trace

xapi/info: replace xapi_fail_intent with xapi_info_unstage infos

Rather than explicitly teeing up for a fail call with xapi_fail_intent, allow
for infos to be unstaged in the event that fail will not in fact be called.

This allows for supplying infos with an assertion in a declarative way

+++ libxlinux/xgrp/xgrp.c
@@ -39,7 +39,7 @@ API xapi uxgetgrgid_r(gid_t gid, struct group * grp, char * buf, size_t buflen, 
   }
 
 finally :
-  xapi_infof("gid", "%zu", gid);
+  xapi_infof("gid", "%lu", (unsigned long)gid);
 coda;
 }
